// src/deploy-in-blockchain/dto/create-deploy-in-blockchain.dto.ts
export class CreateDeployInBlockchainDto {
  vendor: string;
  supplier: string;
  vendorsharepercentage: number;
  suppliersharepercentage: number;
  productprice: number;
  productname: string;
  numberofproducts: number;
  contractdays: number;
}

// {
//     "vendor":"0x6B98F4aE7b9BC8BAAC78E23D69cccDFD037FC035",
//     "supplier":"0x52CaE8F4566F2674423f184E9CA4B2cDDbb95964",
//     "vendorsharepercentage":70,
//     "suppliersharepercentage":30,
//     "productprice":3800000000000000,
//     "productname":"iphone15",
//     "numberofproducts":2,
//     "contractdays":5
// }